Use git am for the patchqueue
I found this trick on the internet: make your patch queue something
suitable for use with "git am".  This makes it much easier to handle
rebasing: You just "git am" onto the current release, "git rebase"
onto the new release, and "git format-patch" to generate the new
patchqueue.

It also removes the need to mess around with PatchNN and %patchNN in
the spec file every time you want to make a change.

I tried using this trick to rebase from Xen 4.2.4 to Xen 4.4.0, and it
worked pretty well.

At the moment, only patches which modify code in the main Xen git repo
can be applied this way.  This does not include either version of qemu
or the out-of-tree blktap 2.5.

Make the .am file with the following command:

git format-patch --stdout -N ${release-tag} > ${path-to-rpm-repo}/SOURCES/xen-queue.am

Signed-off-by: George Dunlap <george.dunlap@eu.citrix.com>